+++ src/lib.rs
@@ -8,17 +8,23 @@
 //! This crate does **not** build Ipopt itself.
 //! The `build.rs` script will generate the bindings based on the system's installed
 //! version of Ipopt.
-//!
-//! This crate only provides access to the raw C bindings of Ipopt.
-//! It is unlikely that this will be nice to use directly.
-//! For more information, please refer to the
-//! [Ipopt C API documentation](https://coin-or.github.io/Ipopt/INTERFACES.html#INTERFACE_C).
-//! For a full Rust usage example, please refer to the
-//! [examples in the project GitHub repository](https://github.com/MattBolitho/ipopt_bindgen/blob/main/examples/hs071.rs).
 
-#![allow(non_upper_case_globals)]
-#![allow(non_camel_case_types)]
-#![allow(non_snake_case)]
+#[cfg(feature = "rust-interface")]
+pub mod application;
+#[cfg(feature = "rust-interface")]
+pub mod results;
+pub mod c_interface;
+#[cfg(feature = "rust-interface")]
+pub mod tnlp;
 
-#[cfg(not(doc))]
-include!(concat!(env!("OUT_DIR"), "/bindings.rs"));
+// Export everything into the root of the crate (seeing as there aren't many public symbols in this
+// library). We *don't* do this for the C interface because the symbols are being automatically
+// generated and ideally consumers will not use them directly anyway, save for creating their own
+// higher-level interface.
+
+#[cfg(feature = "rust-interface")]
+pub use application::*;
+#[cfg(feature = "rust-interface")]
+pub use results::*;
+#[cfg(feature = "rust-interface")]
+pub use tnlp::*;